refactor: 패키지 이동, 클래스 명 개선
- 기존 스타일 참고
